refactor: display functions and tests
I noticed the display and tests file were getting very big, and I decided to refactor this pkg.

This was will become easier to spot those new functions that are not being tested (I spotted some doing this refactor).
