+++ cmd/meroxa/root/apps/describe.go
@@ -20,9 +20,10 @@ import (
 	"context"
 	"errors"
 
+	"github.com/meroxa/cli/utils/display"
+
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
-	"github.com/meroxa/cli/utils"
 	"github.com/meroxa/meroxa-go/pkg/meroxa"
 )
 
@@ -96,7 +97,7 @@ func (d *Describe) Execute(ctx context.Context) error {
 		functions = append(functions, function)
 	}
 
-	output = utils.AppTable(app, resources, connectors, functions)
+	output = display.AppTable(app, resources, connectors, functions)
 
 	d.logger.Info(ctx, output)
 	d.logger.JSON(ctx, app)
