+++ cmd/meroxa/root/apps/list.go
@@ -19,9 +19,10 @@ package apps
 import (
 	"context"
 
+	"github.com/meroxa/cli/utils/display"
+
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
-	"github.com/meroxa/cli/utils"
 	"github.com/meroxa/meroxa-go/pkg/meroxa"
 )
 
@@ -67,7 +68,7 @@ func (l *List) Execute(ctx context.Context) error {
 	}
 
 	l.logger.JSON(ctx, apps)
-	l.logger.Info(ctx, utils.AppsTable(apps, l.hideHeaders))
+	l.logger.Info(ctx, display.AppsTable(apps, l.hideHeaders))
 
 	return nil
 }
