+++ cmd/meroxa/root/connectors/describe.go
@@ -20,7 +20,7 @@ import (
 	"context"
 	"errors"
 
-	"github.com/meroxa/cli/utils"
+	"github.com/meroxa/cli/utils/display"
 
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
@@ -65,7 +65,7 @@ func (d *Describe) Execute(ctx context.Context) error {
 		return err
 	}
 
-	d.logger.Info(ctx, utils.ConnectorTable(connector))
+	d.logger.Info(ctx, display.ConnectorTable(connector))
 	d.logger.JSON(ctx, connector)
 
 	return nil
