+++ cmd/meroxa/root/environments/create.go
@@ -19,11 +19,12 @@ import (
 	"fmt"
 	"strings"
 
+	"github.com/meroxa/cli/utils/display"
+
 	"github.com/manifoldco/promptui"
 
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
-	"github.com/meroxa/cli/utils"
 	"github.com/meroxa/meroxa-go/pkg/meroxa"
 )
 
@@ -134,7 +135,7 @@ func (c *Create) Execute(ctx context.Context) error {
 	}
 
 	if environment.Status.State != meroxa.EnvironmentStatePreflightSuccess {
-		details := utils.EnvironmentPreflightTable(environment)
+		details := display.EnvironmentPreflightTable(environment)
 		c.logger.Errorf(ctx,
 			"Environment %q could not be provisioned because it failed the preflight checks\n%s\n",
 			environment.Name,
