+++ cmd/meroxa/root/environments/list.go
@@ -19,9 +19,10 @@ package environments
 import (
 	"context"
 
+	"github.com/meroxa/cli/utils/display"
+
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
-	"github.com/meroxa/cli/utils"
 	"github.com/meroxa/meroxa-go/pkg/meroxa"
 )
 
@@ -66,7 +67,7 @@ func (l *List) Execute(ctx context.Context) error {
 	}
 
 	l.logger.JSON(ctx, environments)
-	l.logger.Info(ctx, utils.EnvironmentsTable(environments, l.hideHeaders))
+	l.logger.Info(ctx, display.EnvironmentsTable(environments, l.hideHeaders))
 
 	return nil
 }
