+++ cmd/meroxa/root/environments/repair.go
@@ -17,9 +17,10 @@ import (
 	"context"
 	"errors"
 
+	"github.com/meroxa/cli/utils/display"
+
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
-	"github.com/meroxa/cli/utils"
 	"github.com/meroxa/meroxa-go/pkg/meroxa"
 )
 
@@ -78,7 +79,7 @@ func (r *Repair) Execute(ctx context.Context) error {
 	}
 
 	if environment.Status.State != meroxa.EnvironmentStatePreflightSuccess {
-		details := utils.EnvironmentPreflightTable(environment)
+		details := display.EnvironmentPreflightTable(environment)
 		r.logger.Errorf(ctx,
 			"Environment %q could not be repaired because it failed the preflight checks\n%s\n",
 			environment.Name,
