+++ cmd/meroxa/root/environments/update.go
@@ -20,11 +20,12 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/meroxa/cli/utils/display"
+
 	"github.com/manifoldco/promptui"
 
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
-	"github.com/meroxa/cli/utils"
 	"github.com/meroxa/meroxa-go/pkg/meroxa"
 )
 
@@ -112,7 +113,7 @@ func (c *Update) Execute(ctx context.Context) error {
 	}
 
 	if environment.Status.State != meroxa.EnvironmentStatePreflightSuccess {
-		details := utils.EnvironmentPreflightTable(environment)
+		details := display.EnvironmentPreflightTable(environment)
 		c.logger.Errorf(ctx,
 			"Environment %q could not be updated because it failed the preflight checks\n%s\n",
 			environment.Name,
