+++ cmd/meroxa/root/functions/describe.go
@@ -4,9 +4,10 @@ import (
 	"context"
 	"errors"
 
+	"github.com/meroxa/cli/utils/display"
+
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
-	"github.com/meroxa/cli/utils"
 	"github.com/meroxa/meroxa-go/pkg/meroxa"
 )
 
@@ -47,7 +48,7 @@ func (d *Describe) Execute(ctx context.Context) error {
 		return err
 	}
 
-	d.logger.Info(ctx, utils.FunctionTable(fun))
+	d.logger.Info(ctx, display.FunctionTable(fun))
 	d.logger.JSON(ctx, fun)
 
 	return nil
