+++ cmd/meroxa/root/functions/list.go
@@ -3,9 +3,10 @@ package functions
 import (
 	"context"
 
+	"github.com/meroxa/cli/utils/display"
+
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
-	"github.com/meroxa/cli/utils"
 	"github.com/meroxa/meroxa-go/pkg/meroxa"
 )
 
@@ -35,7 +36,7 @@ func (l *List) Execute(ctx context.Context) error {
 	}
 
 	l.logger.JSON(ctx, funs)
-	l.logger.Info(ctx, utils.FunctionsTable(funs, l.hideHeaders))
+	l.logger.Info(ctx, display.FunctionsTable(funs, l.hideHeaders))
 
 	return nil
 }
