+++ cmd/meroxa/root/pipelines/describe.go
@@ -20,9 +20,10 @@ import (
 	"context"
 	"errors"
 
+	"github.com/meroxa/cli/utils/display"
+
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
-	"github.com/meroxa/cli/utils"
 	"github.com/meroxa/meroxa-go/pkg/meroxa"
 )
 
@@ -64,7 +65,7 @@ func (d *Describe) Execute(ctx context.Context) error {
 		return err
 	}
 
-	d.logger.Info(ctx, utils.PipelineTable(p))
+	d.logger.Info(ctx, display.PipelineTable(p))
 
 	d.logger.JSON(ctx, p)
