+++ cmd/meroxa/root/transforms/list.go
@@ -19,9 +19,10 @@ package transforms
 import (
 	"context"
 
+	"github.com/meroxa/cli/utils/display"
+
 	"github.com/meroxa/cli/cmd/meroxa/builder"
 	"github.com/meroxa/cli/log"
-	"github.com/meroxa/cli/utils"
 	"github.com/meroxa/meroxa-go/pkg/meroxa"
 )
 
@@ -66,7 +67,7 @@ func (l *List) Execute(ctx context.Context) error {
 	}
 
 	l.logger.JSON(ctx, transforms)
-	l.logger.Info(ctx, utils.TransformsTable(transforms, l.hideHeaders))
+	l.logger.Info(ctx, display.TransformsTable(transforms, l.hideHeaders))
 
 	return nil
 }
