+++ utils/display/resources.go
@@ -1,0 +1,185 @@
+package display
+
+import (
+	"fmt"
+
+	"github.com/alexeyco/simpletable"
+	"github.com/meroxa/meroxa-go/pkg/meroxa"
+)
+
+func ResourceTable(res *meroxa.Resource) string {
+	tunnel := "N/A"
+	if res.SSHTunnel != nil {
+		tunnel = "SSH"
+	}
+
+	mainTable := simpletable.New()
+	mainTable.Body.Cells = [][]*simpletable.Cell{
+		{
+			{Align: simpletable.AlignRight, Text: "UUID:"},
+			{Text: res.UUID},
+		},
+		{
+			{Align: simpletable.AlignRight, Text: "Name:"},
+			{Text: res.Name},
+		},
+		{
+			{Align: simpletable.AlignRight, Text: "Type:"},
+			{Text: string(res.Type)},
+		},
+		{
+			{Align: simpletable.AlignRight, Text: "URL:"},
+			{Text: res.URL},
+		},
+		{
+			{Align: simpletable.AlignRight, Text: "Tunnel:"},
+			{Text: tunnel},
+		},
+		{
+			{Align: simpletable.AlignRight, Text: "State:"},
+			{Text: string(res.Status.State)},
+		},
+	}
+
+	if res.Status.Details != "" {
+		mainTable.Body.Cells = append(mainTable.Body.Cells, []*simpletable.Cell{
+			{Align: simpletable.AlignRight, Text: "State details:"},
+			{Text: res.Status.Details},
+		})
+	}
+
+	if res.Environment != nil {
+		if res.Environment.UUID.Valid {
+			mainTable.Body.Cells = append(mainTable.Body.Cells, []*simpletable.Cell{
+				{Align: simpletable.AlignRight, Text: "Environment UUID:"},
+				{Text: res.Environment.UUID.String},
+			})
+		}
+
+		if res.Environment.Name.Valid {
+			mainTable.Body.Cells = append(mainTable.Body.Cells, []*simpletable.Cell{
+				{Align: simpletable.AlignRight, Text: "Environment Name:"},
+				{Text: res.Environment.Name.String},
+			})
+		}
+	} else {
+		mainTable.Body.Cells = append(mainTable.Body.Cells, []*simpletable.Cell{
+			{Align: simpletable.AlignRight, Text: "Environment Name:"},
+			{Text: string(meroxa.EnvironmentTypeCommon)},
+		})
+	}
+
+	mainTable.SetStyle(simpletable.StyleCompact)
+
+	return mainTable.String()
+}
+
+func ResourcesTable(resources []*meroxa.Resource, hideHeaders bool) string {
+	if len(resources) != 0 {
+		table := simpletable.New()
+
+		if !hideHeaders {
+			table.Header = &simpletable.Header{
+				Cells: []*simpletable.Cell{
+					{Align: simpletable.AlignCenter, Text: "UUID"},
+					{Align: simpletable.AlignCenter, Text: "NAME"},
+					{Align: simpletable.AlignCenter, Text: "TYPE"},
+					{Align: simpletable.AlignCenter, Text: "ENVIRONMENT"},
+					{Align: simpletable.AlignCenter, Text: "URL"},
+					{Align: simpletable.AlignCenter, Text: "TUNNEL"},
+					{Align: simpletable.AlignCenter, Text: "STATE"},
+				},
+			}
+		}
+
+		for _, res := range resources {
+			tunnel := "N/A"
+			if res.SSHTunnel != nil {
+				tunnel = "SSH"
+			}
+
+			var env string
+
+			if res.Environment != nil && res.Environment.Name.Valid {
+				env = res.Environment.Name.String
+			} else {
+				env = string(meroxa.EnvironmentTypeCommon)
+			}
+
+			r := []*simpletable.Cell{
+				{Align: simpletable.AlignRight, Text: res.UUID},
+				{Text: res.Name},
+				{Text: string(res.Type)},
+				{Text: env},
+				{Text: res.URL},
+				{Align: simpletable.AlignCenter, Text: tunnel},
+				{Align: simpletable.AlignCenter, Text: string(res.Status.State)},
+			}
+
+			table.Body.Cells = append(table.Body.Cells, r)
+		}
+		table.SetStyle(simpletable.StyleCompact)
+		return table.String()
+	}
+
+	return ""
+}
+
+func PrintResourcesTable(resources []*meroxa.Resource, hideHeaders bool) {
+	fmt.Println(ResourcesTable(resources, hideHeaders))
+}
+
+func ResourceTypesTable(types []string, hideHeaders bool) string {
+	table := simpletable.New()
+
+	if !hideHeaders {
+		table.Header = &simpletable.Header{
+			Cells: []*simpletable.Cell{
+				{Align: simpletable.AlignCenter, Text: "TYPES"},
+			},
+		}
+	}
+
+	for _, t := range types {
+		r := []*simpletable.Cell{
+			{Align: simpletable.AlignRight, Text: t},
+		}
+
+		table.Body.Cells = append(table.Body.Cells, r)
+	}
+	table.SetStyle(simpletable.StyleCompact)
+	return table.String()
+}
+
+func PrintResourceTypesTable(types []string, hideHeaders bool) {
+	fmt.Println(ResourceTypesTable(types, hideHeaders))
+}
+
+func extendedResourcesTable(resources []*meroxa.Resource, connectors []*meroxa.Connector) string {
+	if len(resources) == 0 {
+		return ""
+	}
+	subTable := "\tResources\n"
+
+	var r *meroxa.Resource
+	for _, c := range connectors {
+		found := false
+		for _, resource := range resources {
+			if resource.Name == c.ResourceName {
+				r = resource
+				found = true
+				break
+			}
+		}
+		if !found {
+			panic("internal error")
+		}
+
+		subTable += fmt.Sprintf("\t    %s (%s)\n", r.Name, string(c.Type))
+		subTable += fmt.Sprintf("\t\t%5s:   %s\n", "UUID", r.UUID)
+		subTable += fmt.Sprintf("\t\t%5s:   %s\n", "Type", string(r.Type))
+		subTable += fmt.Sprintf("\t\t%5s:   %s\n", "State", string(c.State))
+	}
+
+	return subTable
+}
